[style] 선택 마커 초기화 로직 수정 #36
